Feature/25d (#43)
* refactoring

* Update test_transactions.py

* refactoring

* fixes

* Update test_transfers.py
